### 2021-02-04 #3005 @kyleniemeyer
Hello @vbaliga, this pre-review issue is where we find an editor, reviewers, and also resolve any obvious issues.

Your paper is currently missing an explicit Statement of Need section that illustrates the research purpose of the software, which is one of our [required paper elements](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). Can you please add this?
