### 2021-02-04 #3005 @vbaliga
Hi @kyleniemeyer , thanks for your feedback. I have just updated the paper to make the Statement of Need more clear by making it its own section. This change was made in https://github.com/vbaliga/gaussplotR/commit/e4ef931e1726213b776b6c1112d66173165c9f12

Please let me know if there's anything else you'd like me to address. And thanks for getting this submission into the queue for consideration!
