### 2021-02-08 #3005 @becarioprecario
Thanks for the invitation, but I am not available for making reviews at the moment.

El 8 feb 2021, a las 20:41, Kristina Riemer <notifications@github.com<mailto:notifications@github.com>> escribió:

@becarioprecario<https://github.com/becarioprecario> @emilyriederer<https://github.com/emilyriederer> @cddesja<https://github.com/cddesja> would any of you be able and willing to review this JOSS submission?
