### 2021-02-10 #3005 @cddesja
Sure, I can do this.

On Mon, Feb 8, 2021 at 2:41 PM Kristina Riemer <notifications@github.com>
wrote:

> @becarioprecario <https://github.com/becarioprecario> @emilyriederer
> <https://github.com/emilyriederer> @cddesja <https://github.com/cddesja>
> would any of you be able and willing to review this JOSS submission?
>
