### 2021-02-13 0d456c9
small correction to z_diff
correction to z_diff calculation within characterize_gaussian_fits()
