+++ R/characterize_gaussian_fits.R
@@ -349,8 +349,8 @@ Please revise 'fit_objects_list'."
 
   ## Zdiff
   Z_diff <-
-    ((R_ind - R_spd) / ((1 / (length(original) - 3)) +
-                          (1 / (length(original) - 3)))) * 1 / 2
+    (ZF_ind - ZF_spd) / ((1 / (length(original) - 3)) +
+                           (1 / (length(original) - 3))) ^ 0.5
 
 
   #### Put it all together and return ####
