### 2021-03-01 #3005 @KristinaRiemer
Thank you for following up @becarioprecario and @emilyriederer, and thanks @cddesja for offering to be a reviewer!

I'm still looking for a second reviewer for this JOSS submission then. Would @brunaw, @daviddewhurst, or @rchew be able and willing to review?
