### 2021-03-01 #3074 @whedon
[REVIEW]: gaussplotR: Fit, Predict and Plot 2D Gaussians in R
**Submitting author:** @vbaliga (<a href="http://orcid.org/0000-0002-9367-8974">Vikram B. Baliga</a>)
**Repository:** <a href="https://github.com/vbaliga/gaussplotR" target ="_blank">https://github.com/vbaliga/gaussplotR</a>
**Version:** v0.2.5
