### 2021-03-01 #3074 @brunaw
I'm done with my review, a few comments justifying what I missed in the package & suggestions are below:

- I couldn't straightforwardly install the package from either CRAN or GitHub because of the R version requirement (>= 4.0.0). I find this a high requirement given that this version (4.0.0) is not even 1 year old yet. I managed to install the package only by manually changing this requirement and no harm was done with that, so unless the authors really need this requirement I would suggest downgrading it.

- Just for aesthetics purposes, I suggest rounding the numbers when showing them in the summaries of the models.  Using 5 decimals for each value shown used quite a lot of space on my screen.

 - I found no explicit contribution guide, either in the webpage or GitHub repository of the package. This could be added at the end of the README in the GitHub repository.

- About the paper:
     - Assuming the summary to be the main thing that people read, it certainly needs more details. There's just a short statement of what the package does with no 'clear description of the high-level functionality and purpose of the software for a diverse audience'. I suggest giving at least a quick explanation on 2D Gaussians, why we use them, what are a few examples of it, and why/how the package helps you fitting/plotting them.
   - The same goes for the statement of need, there's seem to be very little justification of why we need this package (I believe we do need it, but the justification for that is lacking), and what is already written seem to be quite generic.
   - References need to be fixed in the statement of need.
   - More references can be added, which will both help to improve the statement of need (by showing how 2D Gaussians have been used in the literature) and to improve the main summary itself.
    - Even though the authors state that there's no other package in the literature that automatically handles the fitting of 2D Gaussians, this package itself is based on functions from previous packages (from the `nls` package for instance, and classical 2D plotting functions). I suggest something is written about that and what are the advantages of using this package (besides being quicker & easier) than other such approaches.

Please let me know if any of this isn't clear or if I can improve my suggestions,

Bruna
