### 2021-03-05 #3074 @KristinaRiemer
Thanks for the quick turnaround time on the review @brunaw! Your comments are useful, and I definitely agree with adding contribution information and fleshing out the use cases of the package in the paper. Hopefully @vbaliga will have a chance to address them soon!
