### 2021-03-05 #3074 @cddesja
I will try to do my review this weekend or early next week.

On Fri, Mar 5, 2021 at 10:15 AM Kristina Riemer <notifications@github.com>
wrote:

> Thanks for the quick turnaround time on the review @brunaw
> <https://github.com/brunaw>! Your comments are useful, and I definitely
> agree with adding contribution information and fleshing out the use cases
> of the package in the paper. Hopefully @vbaliga
> <https://github.com/vbaliga> will have a chance to address them soon!
>
