### 2021-03-16 #3074 @cddesja
I am done with my review as well.

- A couple of references need to be fixed, line 14 and line 30.
- The last sentence of the paper you may want to change the word "the" to "an"
- As @brunaw stated there's no contribution guide.
- I do suspect there are some 2d Gaussian packages in R that could be cited as relevant, if not providing, similar functionality. A quick search finds package with similar names.
- I think it would be beneficial to list all the package requirements on your GitHub site.
- I really appreciate the way that the package allows for easy use with ggplot2 and lattice.
- I am not sure that I agree with @brunaw about whether you need to change the R requirements to allow for R less than 4. While I totally agree with her rationale, I don't think it's necessarily your responsibility to make sure that the package is backwards compatible. However, if @brunaw feels strongly about this, then I think the request is more than reasonable.
- Regarding @brunaw request about the statement of need and the summary section, would it be possible to walk the reader through an applied application of your package? For example, you reference an in-press paper, could you walk us through the problem you encountered and how your package helped to solve it in non-technical language?

Chris
