### 2021-03-16 #3074 @brunaw
Hello @cddesja and @vbaliga,

After I posted my review, I was thinking about my comment about the `R` version. Indeed it's not "too much" to ask for `R` to be at least 4.0.0, because it's been a while that this version is available so maybe it's my obligation as the package user to update my `R` already.

So in summary, I think this is up to the authors to decide on. I would personally not request a certain `R` version if my package really didn't depend on it (to avoided issues like I had), but I also don't see any major issues with keeping the version request as it is 😊
