### 2021-03-16 #3074 @cddesja
"I would personally not request a certain R version if my package really didn't depend on it (to avoided issues like I had)"

This is a good reason to not require 4.0. So, I think @brunaw you're right that this requirement should be removed.
