### 2021-03-18 969f1e3
start taking care of JOSS review items
- changed R requirement to 3.3.0 from 4.0.0
- added print and summary methods to round printing output
