+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(print,gaussplotR_fit)
+S3method(summary,gaussplotR_fit)
 export(autofit_gaussian_2D)
 export(characterize_gaussian_fits)
 export(compare_gaussian_fits)
