### 2021-03-19 0e5d69f
edit paper and add contribution guide
