+++ paper/paper.md
@@ -1,46 +1,72 @@
 ---
-title: 'gaussplotR: Fit, Predict and Plot 2D Gaussians in R'
-tags:
-  - R
-  - 2D Gaussian
-  - Gaussian fit
-  - Gaussian orientation
-  - nonlinear least squares
+title: 'gaussplotR: Fit, Predict and Plot 2D-Gaussians in R'
 authors:
-  - name: Vikram B. Baliga
-    orcid: 0000-0002-9367-8974
-    affiliation: 1
-affiliations:
- - name: Department of Zoology, University of British Columbia, Vancouver, British Colombia, Canada V6T 1Z4
-   index: 1
-date: 03 February 2021
+- affiliation: 1
+  name: Vikram B. Baliga
+  orcid: 0000-0002-9367-8974
+date: "03 February 2021"
 bibliography: paper.bib
-
+tags:
+- R
+- 2D-Gaussian
+- Gaussian fit
+- Gaussian orientation
+- nonlinear least squares
+affiliations:
+- index: 1
+  name: Department of Zoology, University of British Columbia, Vancouver, British
+    Colombia, Canada V6T 1Z4
 ---
 
 # Summary
 
-`gaussplotR` is an R package that is designed to fit 2D Gaussian equations to
-data, characterize fitted models, predict new data from fitted models, and
-provide visualizations of fitted 2D Gaussian surfaces. Such tools enable
-researchers to fit models to data that characterize a response (i.e. the height
-of the Gaussian) for a given set of bivariate inputs (i.e. a set of "X" and "Y"
-variables). 
+Should the need to model the relationship between bivariate data and a response
+variable arise, two-dimensional (2D) Gaussian models are often the most
+appropriate choice. For example, Priebe et. al., [@Priebe2003] characterized
+motion-sensitive neurons by fitting 2D-Gaussian functions to neurons' response
+rates as spatial and temporal frequencies of visual stimuli were varied. The
+width and orientation of these fitted 2D-Gaussian surfaces provides insight on
+whether a neuron is "tuned" to particular spatial or temporal domains.
+Two-dimensional Gaussians are also used in other scientific disciplines such as
+physics [@Wu1998; @Kravtsov2004], materials sciences [@Riekel1999], and image
+processing [@Hanuman2013; @Ketenci2013], particularly in medical imaging
+[@Wu2019; @Qadir2020].
+
+Fitting 2D-Gaussian models to data is not always a straightforward process, as
+finding appropriate values for the model's parameters relies on complex
+procedures such as non-linear least-squares. `gaussplotR` is an R package that
+is designed to fit 2D-Gaussian surfaces to data. Should a user supply bivariate
+data (i.e., x-values and y-values) along with a univariate response variable,
+functions within `gaussplotR` will allow for the automatic fitting of a
+2D-Gaussian model to the data. Fitting the model then enables the user to
+characterize various properties of the Gaussian surface (e.g., computing the
+total volume under the surface). Further, new data can be predicted from models
+fit via `gaussplotR`, which in combination with the package's plotting
+functions, can enable smoother-looking plots from relatively sparse input data.
+In principle, tools within `gaussplotR` have broad applicability to a variety of
+scientific disciplines.
 
 
 # Statement of Need
 
-In principle, tools within `gaussplotR` have broad applicability to
-a variety of scientific disciplines. It is not uncommon for researchers to 
-desire to fit a 2D Gaussian to bivariate input data. For example, 
-`@Priebe2003` characterize motion-sensitive neurons by fitting elliptical 2D 
-Gaussian functions to neurons' response rates to varied visual stimuli. At the
-time of writing, we know of no other packages in the `R` ecosystem that 
[…]
