+++ paper/paper.bib
@@ -14,6 +14,30 @@ volume = {71},
 year = {1994}
 }
 
+@article{Wu1998,
+title={Electromagnetic scattering from a multilayered cylinder arbitrarily located in a Gaussian beam, a new recursive algorithms},
+author={Wu, Zhensen and Guo, Lixin},
+journal={Progress in electromagnetics research},
+volume={18},
+pages={317--333},
+year={1998},
+publisher={EMW Publishing}
+}
+
+@article{Riekel1999,
+title = {Aspects of X-ray diffraction on single spider fibers},
+journal = {International Journal of Biological Macromolecules},
+volume = {24},
+number = {2},
+pages = {179-186},
+year = {1999},
+issn = {0141-8130},
+doi = {https://doi.org/10.1016/S0141-8130(98)00084-1},
+url = {https://www.sciencedirect.com/science/article/pii/S0141813098000841},
+author = {C. Riekel and C. Bränden and C. Craig and C. Ferrero and F. Heidelbach and M. Müller},
+keywords = {X-ray microdiffraction, Synchrotron radiation, Spider silk}
+}
+
 @article{Priebe2003,
 author = {Priebe, Nicholas J and Cassanello, Carlos R and Lisberger, Stephen G},
 doi = {10.1523/JNEUROSCI.23-13-05650.2003},
@@ -30,6 +54,19 @@ volume = {23},
 year = {2003}
 }
 
+@article{Kravtsov2004,
+title = {Description of the 2D Gaussian beam diffraction in a free space in frame of eikonal-based complex geometric optics},
+journal = {Wave Motion},
+volume = {40},
+number = {1},
+pages = {23-27},
+year = {2004},
+issn = {0165-2125},
+doi = {https://doi.org/10.1016/j.wavemoti.2003.12.012},
+url = {https://www.sciencedirect.com/science/article/pii/S0165212503001343},
+author = {Yu.A. Kravtsov and P. Berczynski}
+}
+
 @article{Winship2006,
 author = {Winship, Ian R. and Crowder, Nathan A. and Wylie, Douglas R.W.},
 doi = {10.1152/jn.00921.2005},
@@ -42,4 +79,61 @@ title = {{Quantitative Reassessment of Speed Tuning in the Accessory Optic Syste
 url = {http://www.physiology.org/doi/10.1152/jn.00921.2005},
 volume = {95},
 year = {2006}
+}
+
+@INPROCEEDINGS{Ketenci2013,
+author={S. {Ketenci} and B. {Gencturk}},
+booktitle={Eurocon 2013},
+title={Performance analysis in common color spaces of 2D Gaussian Color Model for skin segmentation},
+year={2013},
+volume={},
+number={},
+pages={1653-1657},
+doi={10.1109/EUROCON.2013.6625198}
+}
+
+@INPROCEEDINGS{Hanuman2013,
+author={M. C. {Hanumantharaju} and M. {Ravishankar} and D. R. {Rameshbabu}},
+booktitle={2013 International Conference on Advances in Computing, Communications and Informatics (ICACCI)},
+title={Design and FPGA implementation of an 2D Gaussian surround function with reduced on-chip memory utilization},
+year={2013},
+volume={},
+number={},
+pages={604-609},
+doi={10.1109/ICACCI.2013.6637241}
+}
+
+@article{Wu2019,
+title = {Hemorrhage detection in fundus image based on 2D Gaussian fitting and human visual characteristics},
+journal = {Optics & Laser Technology},
+volume = {110},
+pages = {69-77},
+year = {2019},
+note = {Special Issue: Optical Imaging for Extreme Environment},
+issn = {0030-3992},
[…]
