### 2021-03-19 #3074 @vbaliga
Hi @brunaw and @cddesja,

Thanks sincerely to both of you for reviewing `gaussplotR` and providing feedback. I think you've both found ways to improve the package and the corresponding paper, and I am grateful for it. Here are my point-by-point replies:

### @brunaw

>     * I couldn't straightforwardly install the package from either CRAN or GitHub because of the R version requirement (>= 4.0.0). I find this a high requirement given that this version (4.0.0) is not even 1 year old yet. I managed to install the package only by manually changing this requirement and no harm was done with that, so unless the authors really need this requirement I would suggest downgrading it.

I agree. Forcing users to have R >= 4.0.0 was too stringent a requirement and not necessary for the package's current functionality. I scanned through all of `gaussplotR`'s dependencies and found that R >= 3.3.0 is a suitable requirement as `rgl` depends on it. Thanks for catching this!

>     * Just for aesthetics purposes, I suggest rounding the numbers when showing them in the summaries of the models.  Using 5 decimals for each value shown used quite a lot of space on my screen.

This was a great suggestion and one that motivated me to write custom `print()` and `summary()` methods for my `fit_gaussian_2D()` outputs. I have now simplified the printed output to be more clean and opted to round each parameter estimate to 2 decimal places.

>     * I found no explicit contribution guide, either in the webpage or GitHub repository of the package. This could be added at the end of the README in the GitHub repository.

Thanks. The reason this wasn't there before was because it doesn't seem to be a requirement for JOSS. But in any case, I agree with the merit of adding it and have placed a section within the README that provides this info. Thanks for the suggestion.

>     * About the paper:
>
>       * Assuming the summary to be the main thing that people read, it certainly needs more details. There's just a short statement of what the package does with no 'clear description of the high-level functionality and purpose of the software for a diverse audience'. I suggest giving at least a quick explanation on 2D Gaussians, why we use them, what are a few examples of it, and why/how the package helps you fitting/plotting them.
>       * The same goes for the statement of need, there's seem to be very little justification of why we need this package (I believe we do need it, but the justification for that is lacking), and what is already written seem to be quite generic.
>       * References need to be fixed in the statement of need.
>       * More references can be added, which will both help to improve the statement of need (by showing how 2D Gaussians have been used in the literature) and to improve the main summary itself.
>       * Even though the authors state that there's no other package in the literature that automatically handles the fitting of 2D Gaussians, this package itself is based on functions from previous packages (from the `nls` package for instance, and classical 2D plotting functions). I suggest something is written about that and what are the advantages of using this package (besides being quicker & easier) than other such approaches.

Thanks for this as well. I have made several changes to the paper, particularly to the summary and statement of need, to give readers a clearer understanding of the basic problems and how `gaussplotR` can provide the tools to solve them. Thanks also for catching the typos re: references and for motivating me to find additional references to showcase the applicability of 2D-Gaussians.

### @cddesja

>     A couple of references need to be fixed, line 14 and line 30.

Thanks for catching this. I believe I have now caught the errors in reference formatting.

>     The last sentence of the paper you may want to change the word "the" to "an"

I've opted to rephrase the last part of the sentence to give a couple more details.

>     As @brunaw stated there's no contribution guide.

Thanks -- as noted above, I have added a section to the README.

>     I do suspect there are some 2d Gaussian packages in R that could be cited as relevant, if not providing, similar functionality. A quick search finds package with similar names.

Thanks, I agree. I have added a citation to the `imagefx` package, given what `imagefx::build.gaus()` is capable of performing. I also note what `stats::nls()` can achieve, though it is not strictly limited to estimating parameters for 2D-Gaussians. Please let me know if you had any other packages in mind -- I have not found any others that are relevant.

>     I think it would be beneficial to list all the package requirements on your GitHub site.

With all due respect, I disagree. This is not a standard practice, and I have only seen it done on the rare occasion that a package has strict requirements (e.g. needing a java installation on the computer, only working on MacOS, etc). Moreover, the requirements for `gaussplotR` are pretty light; there are only four dependencies, two of which are very likely to already be on a user's machine. Please also note that following your & @brunaw's advice, I have relaxed the requirement of R >= 4.0.0, as I agree it is definitely not necessary.

>     I really appreciate the way that the package allows for easy use with ggplot2 and lattice.

Thanks!

>     I am not sure that I agree with @brunaw about whether you need to change the R requirements to allow for R less than 4. While I totally agree with her rationale, I don't think it's necessarily your responsibility to make sure that the package is backwards compatible. However, if @brunaw feels strongly about this, then I think the request is more than reasonable.

As noted above, I am comfortable with relaxing this requirement. It was not necessary and I thank you @brunaw for catching it

>     Regarding @brunaw request about the statement of need and the summary section, would it be possible to walk the reader through an applied application of your package? For example, you reference an in-press paper, could you walk us through the problem you encountered and how your package helped to solve it in non-technical language?

I have made large revisions to the beginning of the paper. From the outset, I have added an explanation of how 2D-Gaussians are used in Priebe et al., 2003 in relatively non-technical language. I would prefer to not go into detail about our in-prep paper, though I will note that much of the analysis in it follow methods used by Priebe et al. So I am hoping that with the revisions I've made to the summary and statement of need, I have now made the motivation behind writing `gaussplotR` and how it is useful more clear to the reader. Does that sound ok?

Thanks again to both of you!

Best regards,
Vikram
🐢
