### 2021-03-19 #3074 @cddesja
@vbaliga, I am fine with your changes. Regarding

> I think it would be beneficial to list all the package requirements on your GitHub site.

I was just basing this off the Installation Instruction checklist. I agree that your package has few requirements and If the user does want to see them they can visit: https://cran.r-project.org/web/packages/gaussplotR/index.html
