### 2021-03-31 #3074 @KristinaRiemer
Hi @vbaliga, the paper looks good except for a couple of smaller errors in the references. Some of the references have "https://doi.org/" listed twice when it should only occur once, and the Wu & Guo references is missing a DOI.

Once those are fixed, you should create a new tagged release of the repo and create an archive of the code (on Zenodo, institutional repo, etc.). Then if you could provide the new version number and the archive DOI here!
