### 2021-03-31 875fbd4
fix DOIs and bump version to 0.2.4
