+++ paper/paper.bib
@@ -21,6 +21,7 @@ journal={Progress in electromagnetics research},
 volume={18},
 pages={317--333},
 year={1998},
+doi = {10.2528/pier97071100},
 publisher={EMW Publishing}
 }
 
@@ -32,7 +33,7 @@ number = {2},
 pages = {179-186},
 year = {1999},
 issn = {0141-8130},
-doi = {https://doi.org/10.1016/S0141-8130(98)00084-1},
+doi = {10.1016/S0141-8130(98)00084-1},
 url = {https://www.sciencedirect.com/science/article/pii/S0141813098000841},
 author = {C. Riekel and C. Bränden and C. Craig and C. Ferrero and F. Heidelbach and M. Müller},
 keywords = {X-ray microdiffraction, Synchrotron radiation, Spider silk}
@@ -62,7 +63,7 @@ number = {1},
 pages = {23-27},
 year = {2004},
 issn = {0165-2125},
-doi = {https://doi.org/10.1016/j.wavemoti.2003.12.012},
+doi = {10.1016/j.wavemoti.2003.12.012},
 url = {https://www.sciencedirect.com/science/article/pii/S0165212503001343},
 author = {Yu.A. Kravtsov and P. Berczynski}
 }
@@ -111,7 +112,7 @@ pages = {69-77},
 year = {2019},
 note = {Special Issue: Optical Imaging for Extreme Environment},
 issn = {0030-3992},
-doi = {https://doi.org/10.1016/j.optlastec.2018.07.049},
+doi = {10.1016/j.optlastec.2018.07.049},
 url = {https://www.sciencedirect.com/science/article/pii/S0030399218308508},
 author = {Jun Wu and Shihao Zhang and Zhitao Xiao and Fang Zhang and Lei Geng and Shiliang Lou and Mengjia Liu}
 }
@@ -124,7 +125,7 @@ volume = {68},
 pages = {101897},
 year = {2021},
 issn = {1361-8415},
-doi = {https://doi.org/10.1016/j.media.2020.101897},
+doi = {10.1016/j.media.2020.101897},
 url = {https://www.sciencedirect.com/science/article/pii/S1361841520302619},
 author = {Hemin Ali Qadir and Younghak Shin and Johannes Solhusvik and Jacob Bergsland and Lars Aabakken and Ilangko Balasingham}
 }
