+++ README.knit.md
@@ -1,60 +1,62 @@
+---
+output: github_document
+---
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+
+
 
 # gaussplotR <img src='man/figures/logo.png' align="right" height="138.5" />
 
 <!-- badges: start -->
-
-[![Project Status: Active – The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R build
-status](https://github.com/vbaliga/gaussplotR/workflows/R-CMD-check/badge.svg)](https://github.com/vbaliga/gaussplotR/actions)
-[![Codecov test
-coverage](https://codecov.io/gh/vbaliga/gaussplotR/graph/badge.svg)](https://codecov.io/gh/vbaliga/gaussplotR?branch=master)  
+[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![R build status](https://github.com/vbaliga/gaussplotR/workflows/R-CMD-check/badge.svg)](https://github.com/vbaliga/gaussplotR/actions)
+[![Codecov test coverage](https://codecov.io/gh/vbaliga/gaussplotR/graph/badge.svg)](https://codecov.io/gh/vbaliga/gaussplotR?branch=master)  
 [![status](https://joss.theoj.org/papers/685673937faee896ddd4a48d825280cf/status.svg)](https://joss.theoj.org/papers/685673937faee896ddd4a48d825280cf)  
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4041073.svg)](https://doi.org/10.5281/zenodo.4041073)
-[![CRAN
-status](https://www.r-pkg.org/badges/version/gaussplotR)](https://CRAN.R-project.org/package=gaussplotR)
+[![CRAN status](https://www.r-pkg.org/badges/version/gaussplotR)](https://CRAN.R-project.org/package=gaussplotR)
 <!-- badges: end -->
 
-`gaussplotR` provides functions to fit two-dimensional Gaussian
-functions, predict values from such functions, and produce plots of
-predicted data.
+`gaussplotR` provides functions to fit two-dimensional Gaussian functions,
+predict values from such functions, and produce plots of predicted data.
 
 ## Installation
 
 You can install `gaussplotR` from CRAN via:
 
-``` r
+
+```r
 install.packages("gaussplotR")
 ```
 
 Or to get the latest (developmental) version through github, use:
+  
 
-``` r
+```r
 devtools::install_github("vbaliga/gaussplotR")
 ```
 
+
 ## Example
 
-The function `fit_gaussian_2D()` is the workhorse of `gaussplotR`. It
-uses `stats::nls()` to find the best-fitting parameters of a 2D-Gaussian
-fit to supplied data based on one of three formula choices. The function
-`autofit_gaussian_2D()` can be used to automatically figure out the best
-formula choice and arrive at the best-fitting parameters.
+The function `fit_gaussian_2D()` is the workhorse of `gaussplotR`. It uses
+`stats::nls()` to find the best-fitting parameters of a 2D-Gaussian fit to
+supplied data based on one of three formula choices. The function
+`autofit_gaussian_2D()` can be used to automatically figure out the best formula
+choice and arrive at the best-fitting parameters.
 
-The `predict_gaussian_2D()` function can then be used to predict values
-from the Gaussian over a supplied grid of X- and Y-values (generated
-here via `expand.grid()`). This is useful if the original data is
-relatively sparse and interpolation of values is desired.
[…]
