### 2021-03-31 45e7230
remove "file" field from paper bib
