+++ paper/paper.bib
@@ -1,7 +1,6 @@
 @article{Levitt1994,
 author = {Levitt, J. B. and Kiper, D. C. and Movshon, J. A.},
 doi = {10.1152/jn.1994.71.6.2517},
-file = {:Users/thalassoma/Dropbox/mendeley{\_}library/Levitt, Kiper, Movshon/1994/Levitt, Kiper, Movshon - 1994 - Journal of Neurophysiology.pdf:pdf},
 issn = {0022-3077},
 journal = {Journal of Neurophysiology},
 month = {jun},
@@ -42,7 +41,6 @@ keywords = {X-ray microdiffraction, Synchrotron radiation, Spider silk}
 @article{Priebe2003,
 author = {Priebe, Nicholas J and Cassanello, Carlos R and Lisberger, Stephen G},
 doi = {10.1523/JNEUROSCI.23-13-05650.2003},
-file = {:Users/thalassoma/Dropbox/mendeley{\_}library/Priebe, Cassanello, Lisberger/2003/Priebe, Cassanello, Lisberger - 2003 - The Journal of Neuroscience.pdf:pdf},
 issn = {0270-6474},
 journal = {The Journal of Neuroscience},
 keywords = {direction tuning,mt,spatial frequency,speed tuning,visual cortex,visual motion processing},
