+++ README.knit.md
@@ -1,191 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-
-
-# gaussplotR <img src='man/figures/logo.png' align="right" height="138.5" />
-
-<!-- badges: start -->
-[![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![R build status](https://github.com/vbaliga/gaussplotR/workflows/R-CMD-check/badge.svg)](https://github.com/vbaliga/gaussplotR/actions)
-[![Codecov test coverage](https://codecov.io/gh/vbaliga/gaussplotR/graph/badge.svg)](https://codecov.io/gh/vbaliga/gaussplotR?branch=master)  
-[![status](https://joss.theoj.org/papers/685673937faee896ddd4a48d825280cf/status.svg)](https://joss.theoj.org/papers/685673937faee896ddd4a48d825280cf)  
-[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.4041073.svg)](https://doi.org/10.5281/zenodo.4041073)
-[![CRAN status](https://www.r-pkg.org/badges/version/gaussplotR)](https://CRAN.R-project.org/package=gaussplotR)
-<!-- badges: end -->
-
-`gaussplotR` provides functions to fit two-dimensional Gaussian functions,
-predict values from such functions, and produce plots of predicted data.
-
-## Installation
-
-You can install `gaussplotR` from CRAN via:
-
-
-```r
-install.packages("gaussplotR")
-```
-
-Or to get the latest (developmental) version through github, use:
-  
-
-```r
-devtools::install_github("vbaliga/gaussplotR")
-```
-
-
-## Example
-
-The function `fit_gaussian_2D()` is the workhorse of `gaussplotR`. It uses
-`stats::nls()` to find the best-fitting parameters of a 2D-Gaussian fit to
-supplied data based on one of three formula choices. The function
-`autofit_gaussian_2D()` can be used to automatically figure out the best formula
-choice and arrive at the best-fitting parameters.
-
-The `predict_gaussian_2D()` function can then be used to predict values from
-the Gaussian over a supplied grid of X- and Y-values (generated here via 
-`expand.grid()`). This is useful if the original data is relatively sparse and
-interpolation of values is desired.
-
-Plotting can then be achieved via `ggplot_gaussian_2D()`, but note that the 
-`data.frame` created by `predict_gaussian_2D()` can be supplied to other 
-plotting frameworks such as `lattice::levelplot()`. A 3D plot can also be 
-produced via `rgl_gaussian_2D()` (not shown here).
-
-
-```r
-library(gaussplotR)
-
-## Load the sample data set
-data(gaussplot_sample_data)
-
-## The raw data we'd like to use are in columns 1:3
-samp_dat <-
-  gaussplot_sample_data[,1:3]
-
-
-#### Example 1: Unconstrained elliptical ####
-## This fits an unconstrained elliptical by default
-gauss_fit_ue <-
-  fit_gaussian_2D(samp_dat)
-
-## Generate a grid of X- and Y- values on which to predict
-grid <-
-  expand.grid(X_values = seq(from = -5, to = 0, by = 0.1),
-              Y_values = seq(from = -1, to = 4, by = 0.1))
-
-## Predict the values using predict_gaussian_2D
-gauss_data_ue <-
-  predict_gaussian_2D(
-    fit_object = gauss_fit_ue,
-    X_values = grid$X_values,
[…]
