### 2021-03-31 #3074 @vbaliga
Hi @KristinaRiemer, thanks very much. I believe I have taken care of the DOI issues in the paper in the latest commits. I have now also:
* created a new release (v0.2.4, see https://github.com/vbaliga/gaussplotR/releases/tag/v0.2.4)
* triggered a Zenodo version update (see https://doi.org/10.5281/zenodo.4041073 for All Versions; and https://doi.org/10.5281/zenodo.4654036 for v0.2.4 specifically)

Please let me know if there's anything else I should do - thanks again!
