### 2021-04-01 #3074 @KristinaRiemer
Great, thanks for doing that @vbaliga! So we want the archived code to match the submission, so could you change the title on Zenodo to "gaussplotR: Fit, Predict and Plot 2D-Gaussians in R"?
