### 2021-04-01 #3074 @vbaliga
@KristinaRiemer sure thing! I just pushed an update to Zenodo (v0.2.5, see: https://doi.org/10.5281/zenodo.4657750). The corresponding release on the github repo is available here: https://github.com/vbaliga/gaussplotR/releases/tag/v0.2.5
