### 2021-04-06 e0ab0ca
two minor paper changes
