### 2021-04-06 #3074 @danielskatz
👋 @vbaliga - there are two small changes in https://github.com/vbaliga/gaussplotR/pull/1 - if you can merge this, we should be able to proceed to publication.
